+++ js/edit-photo-form.js
@@ -7,14 +7,14 @@ const bodyNode = document.querySelector('body');
 const userCloseElementNode = document.querySelector('.img-upload__cancel');
 const formModalNode = document.querySelector('.img-upload__form');
 
-function openedFormModal () {
+function openEditPhotoFormModal () {
   overlayModalNode.classList.remove('hidden');
   bodyNode.classList.add('modal-open');
 
   document.addEventListener('keydown', onFormModalEscKeydown);
 }
 
-function closedFormModal () {
+function closeEditPhotoFormModal () {
   overlayModalNode.classList.add('hidden');
   bodyNode.classList.remove('modal-open');
   formModalNode.reset();
@@ -25,18 +25,18 @@ function closedFormModal () {
 function onFormModalEscKeydown (evt) {
   if (isEscKey(evt)) {
     evt.preventDefault();
-    closedFormModal();
+    closeEditPhotoFormModal();
   }
 }
 
-formControlNode.addEventListener('change', openedFormModal);
+formControlNode.addEventListener('change', openEditPhotoFormModal);
 formControlNode.addEventListener('keydown', (evt) => {
   if (isEnterKey(evt)) {
-    openedFormModal();
+    openEditPhotoFormModal();
   }
 });
 
-userCloseElementNode.addEventListener('click', closedFormModal);
-userCloseElementNode.addEventListener('keydown', closedFormModal);
+userCloseElementNode.addEventListener('click', closeEditPhotoFormModal);
+userCloseElementNode.addEventListener('keydown', closeEditPhotoFormModal);
 
 export {onFormModalEscKeydown};
