+++ js/main.js
@@ -1,7 +1,7 @@
 import {createPhotoDescription} from './mock/photo-description.js';
 import {renderPhotoPreview} from './render-photo-preview.js';
-import './form.js';
-import './validate-form.js';
+import './edit-photo-form.js';
+import './edit-photo-validate-form.js';
 
 const COUNT_DESCRIPTION = 25;
